chore: clean up desktop app config and justfile
- Fix folder paths in desktop config
- Move suppressQtWarnings to main.py
- Simplify justfile recipes by adding variables for common commands
- Add LOG_LEVEL to env example
