+++ .env.example
@@ -43,3 +43,6 @@ TOKEN_SECRET=any-64-character-ultra-secure-randomly-generated-secret
 # Ngrok configuration (optional)
 NGROK_AUTHTOKEN=
 NGROK_DOMAIN=
+
+# Optional configuration for development and debugging
+LOG_LEVEL=INFO
